requirements.txt
streamlit==1.37.1
python-docx==1.1.2
python-pptx==0.6.23
pypdf==5.0.1
graphviz==0.20.3
pyvis==0.3.2
networkx==3.2.1
pandas==2.2.2
